# chore: optimize uuid.v4 by 4.3x (430%) (internal stringify by 33x)