@LinusU 2021-11-24
The new `safeStringify` no longer validates that the version is between 0-5, or that version 0 consists only of zeros, as the previous `stringify` does.

How about keeping the public `stringify` as it was before, but have an internal `unsafeStringify` that we use internally from `v1`/`v35`/`v4`? Since there we know that the bytes are valid.

That way this won't be a breaking change either, so it can go out in a regular release.