@ctavan 2021-12-02
> @ctavan Do you know why the `bundlewatch` check isn't running here?

The browser tests and bundlewatch both use API keys (for Browserstack and bundlewatch respectively) and therefore don't run on PRs created from branches of forks. They only run on branches of this very repo.

The reason is that it would otherwise be trivial to extract those secrets (just `console.log(process.env.API_KEY)`).

I haven't checked in a while if there is a solution for this problem by now. But until then we can also just merge and observe the checks on `main`.