@LinusU 2021-12-02
Bundlewatch report for `v8.3.2-14-g577c840` (this PR):

```text
PASS ./examples/browser-rollup/dist/v1-size.js: 802B < 1KB (gzip)
PASS ./examples/browser-rollup/dist/v3-size.js: 1.96KB < 2.1KB (gzip)
PASS ./examples/browser-rollup/dist/v4-size.js: 522B < 716B (gzip)
PASS ./examples/browser-rollup/dist/v5-size.js: 1.36KB < 1.5KB (gzip)
PASS ./examples/browser-webpack/dist/v1-size.js: 859B < 1KB (gzip)
PASS ./examples/browser-webpack/dist/v3-size.js: 1.96KB < 2.1KB (gzip)
PASS ./examples/browser-webpack/dist/v4-size.js: 552B < 716B (gzip)
PASS ./examples/browser-webpack/dist/v5-size.js: 1.35KB < 1.5KB (gzip)
```

Bundlewatch report for `v8.3.2-12-gc9e076c` (`main`):

```text
PASS ./examples/browser-rollup/dist/v1-size.js: 912B < 1KB (gzip)
PASS ./examples/browser-rollup/dist/v3-size.js: 1.99KB < 2.1KB (gzip)
PASS ./examples/browser-rollup/dist/v4-size.js: 624B < 716B (gzip)
PASS ./examples/browser-rollup/dist/v5-size.js: 1.39KB < 1.5KB (gzip)
PASS ./examples/browser-webpack/dist/v1-size.js: 905B < 1KB (gzip)
PASS ./examples/browser-webpack/dist/v3-size.js: 1.99KB < 2.1KB (gzip)
PASS ./examples/browser-webpack/dist/v4-size.js: 589B < 716B (gzip)
PASS ./examples/browser-webpack/dist/v5-size.js: 1.39KB < 1.5KB (gzip)
```

Summary:

&nbsp; | v8.3.2-12-gc9e076c | v8.3.2-14-g577c840
----- | ----- | -----
`rollup/v1-size.js` | `912B` | `802B`
`rollup/v3-size.js` | `1.99KB` | `1.96KB`
`rollup/v4-size.js` | `624B` | `522B`
`rollup/v5-size.js` | `1.39KB` | `1.36KB`
`webpack/v1-size.js` | `905B` | `859B`
`webpack/v3-size.js` | `1.99KB` | `1.96KB`
`webpack/v4-size.js` | `589B` | `552B`
`webpack/v5-size.js` | `1.39KB` | `1.35KB`

Lower bundle sizes overall 🙌