@andrejewski 2021-12-21
It would be great to expose this `unsafeStringify` so others can use it on trusted inputs as well; an `export` is used here but I can't import the code; when I do I get:

```
node
> require('uuid/dist/stringify')
Uncaught:
Error [ERR_PACKAGE_PATH_NOT_EXPORTED]: Package subpath './dist/stringify' is not defined by "exports" in ~/my-project/node_modules/uuid/package.json
```

```
> node -v
v15.14.0
```