@broofa 2021-12-24
@andrejewski  What's your use case for this?  Unless `stringify()` is being called 1M's/second the perf gain of `unsafeStringify()` is unlikely matter.