@Rush 2021-12-24
Same as calling uuid 1M times per second :) I even think that the unsafe one should be the default.