@broofa 2021-12-30
>  it is wonky to pay the validation cost twice

It is also "wonky" to take on the readability / maintenance burden that comes with expanding a module's API unless there is a good reason for it.

Look... I understand the desire to make code as fast as possible, but when it comes to performance, it helps to have a certain amount of pragmatism. The nature of the perf benefit that `unsafeStringify()` represents is such that unless you're stringifying UUIDs at a rate of 1M+ UUIDs / minute / CPU core (or thereabouts), the 1-2 seconds of time it saves is just not that interesting.

And my long-held belief is that nobody is actually doing that.  There just aren't any real-world environments where UUIDs are being generated at that volume.

Hence, my question:  What real-world use cases do you have for this?  I.e. Convince me I'm wrong.  Articulate an application for this where the perf gain actually matters.