@broofa 2021-12-30
@andrejewski The more I think about this, the less inclined I am to expose `unsafeStringify()`.  It's generally not a good idea to incorporate these types of perf optimizations into the public API.

Since you already have a UUID class, it might make sense for you to inline your preferred `stringify` implementation in your class.  (Feel free to copy the `unsafeStringify()` code for that if you'd like)