@Rush 2021-12-30
@andrejewski It's a PITA but since there is a lot of opposition to the idea of exposing the optimized version of stringify maybe you can just copy & paste the stringify function to your codebase and be done with it. 

Not sure why we would make a statement that doing stringification a million times per second is only reserved to `uuid` internals and not to other use cases. I can imagine a dozen use-cases which could benefit from it.

Especially since the optimized version is **30x**, yes 30 times faster.